requirements.txt
discord.py
python-decouple
openai==0.18.1